Rework of the project as a whole.
I essentially reworked the project as a whole, especially the back-end which needed some work . This is after a break with the project and realizing that I could essentially do better. This is also why the dates of the files have changed. I also made a more in-depth README.md file for the main directory, plus an INSTALLATION.md file. for details regarding installing and setting up the project, and another README.md file in the focust-back-end directory, which contains information about the REST commands one can write to the back end server via CURL or Postman.

The focust-front-end directory was also made, but it currently only contains a basic TypeScript React.js project made via the node project manager. This will change with further commits.