Modified "INSTALLATION.md" and removed "README.md" from "focust-front-end"
I simply modified the "INSTALLATION.md" file to contain an instruction to modify the "package.json" file if one changed the port number for the back-end server.

I also got rid of the "README.md" file from "focust-front-end", which was the normal "README.md" one would get when creating a React project.