Created "dependencies.sh"
For this commit, I have done the following things:

1) Created two "dependencies.sh" bash script files for both the front-end and back-end servers to handle the installation of dependencies and have docker run it instead of manually installing the dependencies itself. This is to allow more flexibility in terms of how to handle installation issues that Docker cannot provide, like using a loop to keep using "apt-get update --fix-missing" and trying to install a given dependency again until it gets installed successfully.

_______

I wanted to add a new dependency for the back-end to allow me to check the health of the back-end server and use it as a means to see if I can go ahead and start up the front-end server, but including the dependency "spring-boot-starter-actuator" ended up causing problems when building the back-end server, and spent the entire week trying to debug it to no success, but I will still keep working on it. Hopefully in a future commit I can add a health check to the back-end server.